+++ ydb/public/lib/ydb_cli/commands/ydb_workload_import.cpp
@@ -47,14 +47,14 @@ TWorkloadCommandImport::TUploadCommand::TUploadCommand(NYdbWorkload::TWorkloadPa
     , Initializer(initializer)
 {}
 
-int TWorkloadCommandImport::TUploadCommand::DoRun(NYdbWorkload::IWorkloadQueryGenerator& workloadGen, TConfig& config) {
+int TWorkloadCommandImport::TUploadCommand::DoRun(NYdbWorkload::IWorkloadQueryGenerator& /*workloadGen*/, TConfig& /*config*/) {
     auto dataGeneratorList = Initializer->GetBulkInitialData();
     AtomicSet(ErrorsCount, 0);
     InFlightSemaphore = MakeHolder<TFastSemaphore>(UploadParams.MaxInFlight);
     if (UploadParams.FileOutputPath.IsDefined()) {
         Writer = MakeHolder<TFileWriter>(*this);
     } else {
-        Writer = MakeHolder<TDbWriter>(*this, workloadGen, config);
+        Writer = MakeHolder<TDbWriter>(*this);
     }
     for (auto dataGen : dataGeneratorList) {
         TThreadPoolParams params;
@@ -81,21 +81,11 @@ int TWorkloadCommandImport::TUploadCommand::DoRun(NYdbWorkload::IWorkloadQueryGe
 }
 class TWorkloadCommandImport::TUploadCommand::TDbWriter: public IWriter {
 public:
-    TDbWriter(TWorkloadCommandImport::TUploadCommand& owner, NYdbWorkload::IWorkloadQueryGenerator& workloadGen, TConfig& config)
+    TDbWriter(TWorkloadCommandImport::TUploadCommand& owner)
         : IWriter(owner)
     {
         RetrySettings.RetryUndefined(true);
         RetrySettings.MaxRetries(30);
-        for(const auto& path: workloadGen.GetCleanPaths()) {
-            const auto list = NConsoleClient::RecursiveList(*owner.SchemeClient, config.Database + "/" + path.c_str());
-            for (const auto& entry : list.Entries) {
-                if (entry.Type == NScheme::ESchemeEntryType::ColumnTable || entry.Type == NScheme::ESchemeEntryType::Table) {
-                    const auto tableDescr = owner.TableClient->GetSession(NTable::TCreateSessionSettings()).GetValueSync().GetSession().DescribeTable(entry.Name).ExtractValueSync().GetTableDescription();
-                    auto& params = ArrowCsvParams[entry.Name];
-                    params.Columns = tableDescr.GetTableColumns();
-                }
-            }
-        }
     }
 
     TAsyncStatus WriteDataPortion(NYdbWorkload::IBulkDataGenerator::TDataPortionPtr portion) override {
@@ -120,11 +110,11 @@ public:
 private:
     TAsyncStatus WriteCsv(NYdbWorkload::IBulkDataGenerator::TDataPortionPtr portion) {
         const auto* value = std::get_if<NYdbWorkload::IBulkDataGenerator::TDataPortion::TCsv>(&portion->MutableData());
-        const auto* param = MapFindPtr(ArrowCsvParams, portion->GetTable());
-        if (!param) {
-            return NThreading::MakeFuture(TStatus(EStatus::INTERNAL_ERROR, NYql::TIssues({NYql::TIssue("Table does not exist: " + portion->GetTable())})));
+        const auto param = GetCSVParams(portion->GetTable());
+        if (!param.Status.IsSuccess()) {
+            return NThreading::MakeFuture(param.Status);
         }
-        auto arrowCsv = NKikimr::NFormats::TArrowCSVTable::Create(param->Columns, true);
+        auto arrowCsv = NKikimr::NFormats::TArrowCSVTable::Create(param.Columns, true);
         if (!arrowCsv.ok()) {
             return NThreading::MakeFuture(TStatus(EStatus::INTERNAL_ERROR, NYql::TIssues({NYql::TIssue(arrowCsv.status().ToString())})));
         }
@@ -160,11 +150,36 @@ private:
     }
 
     struct TArrowCSVParams {
+        TStatus Status = TStatus(EStatus::SUCCESS, NYql::TIssues());
         TVector<NYdb::NTable::TTableColumn> Columns;
     };
 
+    TArrowCSVParams GetCSVParams(const TString& table) {
+        auto g = Guard(CsvParamsLock);
+        auto result = ArrowCsvParams.emplace(table, TArrowCSVParams());
+        if (result.second) {
+            auto session = Owner.TableClient->GetSession(NTable::TCreateSessionSettings()).ExtractValueSync();
+            if (!session.IsSuccess()) {
+                auto issues = session.GetIssues();
+                issues.AddIssue("Cannot create session");
+                result.first->second.Status = TStatus(session.GetStatus(), std::move(issues));
+            } else {
+                const auto tableDescr = session.GetSession().DescribeTable(table).ExtractValueSync();
+                if (!tableDescr.IsSuccess()) {
+                    auto issues = tableDescr.GetIssues();
+                    issues.AddIssue("Cannot descibe table " + table);
+                    result.first->second.Status = TStatus(tableDescr.GetStatus(), std::move(issues));
+                } else {
+                    result.first->second.Columns = tableDescr.GetTableDescription().GetTableColumns();
+                }
+            }
+        }
+        return result.first->second;
+    }
+
     TMap<TString, TArrowCSVParams> ArrowCsvParams;
     NRetry::TRetryOperationSettings RetrySettings;
+    TAdaptiveLock CsvParamsLock;
 };
 
 class TWorkloadCommandImport::TUploadCommand::TFileWriter: public IWriter {